Handle ref classes with missing fields.
This is a more aggressive workaround for #557. We already have the ability to tolerate missing fields in codegen. So we fall back to that if we're building up a model of a ref class and can't account for all of its fields.

For value classes we still insist that we get a complete picture, because we intend to use the LLVM types to produce GC info.