Return a properly typed pointer from getStaticFieldAddress.
This was causing [unexpected type in load/store primitive] in two CoreCLR tests:
static_passing_class01 and static_passing_struct01. The result of getStaticFieldAddress
is used as a pointer for a constrained virtual call. Dereferencing this pointer for
CORINFO_TYPE_REFANY resulted in the error.

With this fix we are at 100% when compiling currently enabled CoreCLR tests.