Fix the type of refanyval.
refanyval was returning an interior pointer to a byte instead of
an interior pointer to the type given as an operand to the
refanyval instruction. This was causing downstream code to assert.

Fixing this assertion requires excluding a test that is now failing
due to lack of EH support in a method that is now compiled using
LLILC rather than RyuJIT.