LLILC changes needed to merge latest LLVM.
LLVM now include a coff.h file that defines in enumeration types
constants with the same names that Windows defines using
in the scope of the windows #defines. To avoid that we
include related include files early before anyone has includes
windows.h.

Two other small changes made:

1. The TargetOptions.NoFramePointerElim field no longer exists.
For now I just commented out references to it, but we need to
decide if there is something else we need to do to
ensure we get a frame pointer.

2. The parameters of the LLVMBUILDER->CreateCall method changed.
Now need to pass args. We only used this for making NOPs, so
pass empty args.