Support {add,sub}.ovf.un when the result is a pointer.
ECMA-335 supports addition and subtraction with overflow under the rules
given in table III.8. Because overflowing arithmetic in LLVM requires
calling an intrinsic that only accepts integral arguments, supporting
these operations requires converting the incoming operands to an
intermediate integral type, then converting the result to the ultimate
result type.

This also includes a small fix for errors encountered when compiling with
GCC 4.9.2 on Linux.