Implement GC-Table generation
Translate GcInfo from __llvm.stackmaps section to
CoreCLR GCInfo Encoding.

Testing:
1) We can now build and a few tests in a precise-GC environment.
   The correctness of generated GCInfo is verified by:
   a) Examining the GcInfo dump in the debugger (SOS) and comparing
      it against the assembly dump.
   b) Stepping through the debugger and watch it scan through the
      reported GC-Roots.
   c) The execution crashes for the same functions if the liveness
      information is not reported.

2) The execution crashes if we try to build MsCorLib for precise-GC.
There are a few known issues (as noted in the comments) and the
remaining issues need to be fixed via debugging.

3) The standard LLILC testig is still done using conservative-gc
So, no testing impact is expected for the lab. A separate test-leg will
be added to run tests with precise-GC.

Fixes #31