Add ability to specify CoreCLR tests directory
For users who have cloned CoreCLR into a directory named something other
than coreclr, the current checking of the current directory would fail.
This change allows the user to specify where their coreclr\tests
directory can be found, and allows users to run llilc_runtests.py
outside of coreclr\tests. The script will abort if runtest.cmd is not
found in the directory specified as the coreclr\tests path (or the
current working directory if none is specified). We then run runtest
specifying the full path.