Fix unwind info reporting.
Size passed to EEMemoryManager::registerEHFrames
includes the size of a padding UNWIND_CODE entry
so that the number of elements of the UnwindCode array is even.
Jit interface (allocUnwindInfo) expects the size reported to it not to include the padding
entry.

This was causing an assert in crossgen.exe when using llilcjit.