Remove typedefs for jit layers
The jit layers are really an implementation detail, and don't need to be
exposed from LLILCJit.h.  Even in LLILCJit.cpp, they're only used to
declare the layers; this change inlines them into the declarations (using
decltype(PreviousLayer) to avoid excessive repetition of lower layer
types) and removes them entirely.

This avoids needing to make parallel changes in the typedefs and
declarations when adding/changing layers.