Modify SymbolRef getName and getAddress calls for new signature.
The SymbolRef::getName and SymbolRef.getAddress methods
now use ErrorOr<T> return types to return either an
error code or the desired value. The calls to these
were changed to capture the ErrorOr<T> value and
then test for error. If no error then extract the
value and proceed to use it.