Support for ngen-ing of methods with switches and floating point instructions.
Process all relocations in ObjectLoadListener::recordRelocations.
Fix an issue with read-only section alignment: make sure Constant
pools have a 16 bytes alignment.

Create a copy of ObjectLinkingLayer (EEObjectLinkingLayer)so that
we can suppress relocation resolutiotion by LLVM. Lang Hames is
likely to re-design ObjectLinkingLayer in the future to allow more
control over relocations. When that happens we should be able to get rid
of our copy.

Closes #699.
Closes #700.