Implement translation of jmp msil instruction.
jmp is implemented as a call marked with musttail followed by a ret.

If the caller has a secret parameter we need to pass it to the jmp call
because the signature and the calling convention of the caller and
the musttail call must match. To accommodate that a parallel
LLVM change modifies CLR_SecretParameter calling convention
to make the secret parameter the first pointer-sized parameter
(still passed in R10) rather than a parameter passed with a
"CLR_SecretParameter" attribute.

All indirect parameters are passed as they come to the caller, without
copying them to the current frame.

Indirect result parameter is typed as a managed pointer in a jmp call
because we type indirect result parameter for the caller as a
managed pointer.

Closes #192.