Add support for building Roslyn (and other apps) with LLILC.
test/llilc_run.py is a general utility for running applications
with LLILC as the JIT. It has options to control it that all
start with "--llilc-" and there are no short forms. This is
in order to avoid clashing with the options of the
program being invoked.

test/llilc_roslyn_stage.cmd is a batch script that
will create a bootstrap stage directory for Roslyn
from the existing Roslyn binaries. It then builds
Roslyn using those binaries.