GcInfo: Fix Offsets for NGEN
GcInfoEncoder needs to find the distance (offset) of the
Function entry point to the start of the code section.
For this, it calculated the diffrence between the HotCodeBlock
and the function address reported in __llvm_stackmaps.

However, in an Ngen/AOT build, the FunctionEntry address
recorded in LLVM's StackMap are post-relocation offsets.
Therefore, it cannot be compared to the Start of the CodeBlock
allocated during compilation.

This change fixes this bug by using non-relocated code-block
address for both function-entry and code-block-start.

Fixes #743