GCTables: Report StackBasePointer.
CoreCLR requires that we identify whether a method's frame is
FP or SP based. This is inaddition to reporting the reference pointers
for the slots representing liveness. So, this change reports the
StackBasePointer in the GcTable header.

To facilitate the above, the GcInfo encoding functions are
restructured to be per-function routines.

Fixes #758.