Allow int32 to int64 comparisons in branches
LLILC was already allowing int32 to int64 comparisons
for the opcodes that produce a boolean result.
But the code that adjusted comparison arguments was
only partly duplicated for branch comparisons.

This change factors out the common code that
adjusts comparison arguments so that boolean result
comparison and branch comparison are treated consistently.

With this change test
JIT\Regression\CLR-x86-JIT\V1-M13-RTM\b91223,
which was failing with an assertion failure is now
able to compile the ILGEN_0x9c0b37ec.Method_0xa1f09b04
method. The test still fails later, because the tests
correct execution throws an overflow exception which
LLILC is not yet prepared to catch. So this test
is still excluded but its issue number is
changed to #13.

This change also includes two other small changes:

The MSIL dumps were being written directly to stderr
rather than going through dbgs() and I've observed
that this output does not appear when called from
the debugger. To fix this I've changed the
dbprint method to first format the message into a
string and then output that string to dbgs().
Currently we only use dbPrint to write MSIL
dumps. I allowed a buffer size of 200 characters,
but also used the safe string formatting method (vsnprintf).

In the llilc_run.py file, when verbose output is being
produced I modified the log function so that output to
stdout and stderr were tagged to avoid confusion,
as when they have the save destination the duplication
can be confusing.

Resolves #637.