Additional LLILC debugging aids
The changes makes the following improvements to the
LLILC debugging facilities:

1. A make_sln.py script is added which can be used to create
a Visual Studio 2015 solution that can be used to debug
a managed application using LLILC. Optionally the
script can invoke Visual Studio on the created solution.
The handles setting up all the required environment
variables for making LLILC the alternate JIT.

2. A --windbg-and-args option was added to the
llilc_run.py script to enable debugging the
application being run using windbg.

3. In order to ensure that empty values for
LLILC configuration variables are equivalent to
unspecified environment variables, the
code that LLILC uses to check some configuration
variables was refactored to require that
the feature is only enabled if the variable
is both non-null and non-empty.

4. The Debugging.md documentation was updated
to describe these features.

5. The llilc_run.py and make_sln.py were modified so they
run successfully on either Python 2.7 or Python 3.4.