For former #637 tests that throw exceptions, change to issue #13
Some of the failing excluded tests that formerly were excluded
because of issue #637 still fail even though issue #637 has
been fixed, because they throw exceptions. Change the issue
number for these to #13. I verified in the debugger that
these tests do in fact throw an exception (when compiled with
Ryujit on the CoreClr).

In addition this change deletes some exclusions for tests that
no longer exist. Also one test had "13(" as its issue, so I
fixed that so it is "13".