Make sure subprogram addr and function addr match
For communicating debug locals to the CLR EE, we want to make sure that
the starting address for the current function matches that of the
Subprogram DIE. In the case where we have multiple function symbols in
one module (GC Statepoints, funclets), we want to make sure that we are
passing to the EE locals that belong to the actual function (ie, not the
gc.safepoint_poll function). This change extracts the Subprogram DIE's
low pc (start address) and compares it to the symbol's start address. If
the two are not the same, it returns without collecting local debug
info.