Fix #789 and #797
Closes #789. Closes #797.

For #789, when merging stack states, if the types being merged
are int32 and native int, keeps the first type in order to
conform to the ECMA standard. This fixed 8 tests.

For #797, if a switch has case values whose type does not
match the type of the selector, the types of the case values
are adjusted to match the selector type while preserving
their value. This fixed two tests.