Fix issue #633
This fix handles function pointers in calls by
converting the method pointer to a native
int, as suggested by ECMA II.14.5.

Two of the tests were originally classified as #637
but after that was fixed became #633.

With this fix 11 tests pass. 5 tests were found
to no longer exist. 4 tests still fail with
thrown exceptions and are given issue #13.