Map BasicBlocks to EHRegion nodes
Create a Block -> Region map in the first pass by paying attention to
region boundaries as we advance through the IL, and maintain it as new
blocks are added in the second pass by preserving it when splitting blocks
and such.

Expand the EHRegion struct to include fields that will be useful for
building EH flow.

Remove the setupBlockForEH method that is now superfluous.

Add a stub EnterRegion method to GenIR; this is where landing/EH pads will
be generated as the EH implementation comes online.