Make float to int conversion more determined.
The ECMA-335 spec says that for unchecked floating point to
integer conversion, that if there is overflow then the result
is unspecified. However for compatibility with other jits
we make the result more deterministic. If the target
bit size is less than 64 we first convert to i64 where overflow
is less likely. Then we truncate to the target bit size.

Closes #808 and allows test
\JIT\Regression\CLR-x86-JIT\V1-M10\b05617\b05617.cmd
to pass.