Generate new WinEH representation in reader
Change the reader to generate catchpad/cleanuppad instead of landingpad:
 - Catch handlers translate to catchpads with the token for the caught
   exception type as an argument.
 - Finally handlers translate to cleanuppads with no arguments.
 - Fault handlers translate to cleanuppads with a single dummy argument to
   distinguish them from finally handlers.
 - Filter expressions get outlined into separate filter functions, and
   their handlers get translated to catchpads with the filter function as
   an argument.

The IR generated for finally/fault handlers currently fails verification
because the uses of cleanuppads on cleanuprets/cleanupendpads aren't
dominated by those defs; we'll ultimately want a cleanupcall operator, and
the near-term plan is to duplicate out the exceptional path for each
finally in a post-pass, but this duplication is NYI.

The IR generated for the other handler types looks reasonable on some
small testcases.

None of this IR can be lowered to machine code yet, as LLVM's lowering
does not yet handle these opcodes.