Clone finally bodies
The main reader passes generate IR for finally handlers that reflects the
right semantics, but is ill-formed because the `cleanuppad` at the entry
of a finally doesn't dominate its uses at the exits from that finally
(because there is a side-entry for the non-exceptional paths through the
finally).

Add a post-pass (in `readerPostVisit`) that makes a separate clone for the
non-exception version of each finally handler, leaving the `cleanuppad`
def and uses well-formed on the exception-only path.

The long-term plan is to add a `cleanupcall` operator to LLVM that the
non-exceptional paths can use to target the `cleanuppad` directly (and
perform finally cloning as a subsequent optimization in many cases, but
without the need to introduce potentially-explosive cloning for
correctness).  That work is being deferred to simplify/speed-up the
initial WinEH bring-up.

Closes #835.