Disable type caching across methods in ReadyToRun mode.
ReadyToRun mainv2 test has a series of subtests. One of them is TestGrowingStruct.
The scenario is that an app is ReadyToRun-compiled against a version of a struct in a lib.
Then two new fields are added to the struct and the lib is ReadyToRun recompiled.
Then we attempt to run the app without recompiling. What should happen in this case
is that the runtime should detect that the struct layout has changed and the pre-compiled
code for the method in the app is invalid and the method should be JIT-ed.

The reason that wasn't always happening was because we have per-thread caching of types.
Crossgen records structs that the JIT calls getClassSize for and that allows the runtime to do
the detection described above. In this case more than one method needed the info about
that struct but crossgen recorded just the method that happened to call getClassSize on it first.
So if a method was compiled on the same thread as TestGrowingStruct before it and called getClassSize,
crossgen wouldn't record the TestGrowingStruct's dependency on the struct layout.

The fix is not to do type caching across methods in ReadyToRun mode.

Closes #833.