Update llilc_roslyn_stage script
Previously was using c:/Python34, but that is no longer
installed on lab machines. The script that it invokes
will work with either Pythone 3.4 or 2.7 so just use
the Python on the path.

Add buildtype as an argument to llilc_roslyn_stage.cmd
so it will work with debug, checked, or release builds.

Revise for new Roslyn strategy for running on coreclr.

The Roslyn analyzers sometimes throw task cancellation
exceptions. Since LLILC cannot handle these yet,
inhibit their use.