Fix filter function boundaries
The code to move filter blocks from the root function into their filter
function was neglecting to account for point blocks generated in the 2nd
pass, which get inserted at the end of the list rather than in MSIL order.
Fix that for now by searching through the entire list for blocks in a
given filter.