Account for try regions that follow their handlers
It's legal MSIL for a protected region's handler(s) to precede it, but
fgSwitchRegion wasn't expecting this.  Add a property of try regions that
indicates which of the try region itself or its handlers is lexically
first (called its "EntryRegion"), computed during EH Region Tree
construction, which fgSwitchRegion can use to make sure it calls
fgEnterRegion exactly once for each try, and before switching to any
handler of the try.