Enhance llilc_checkpass.py to report failure details.
llilc_checkpass.py has been enhanced to do a more complete
comparison between a target test run and a baseline run.
Reports are provided of the failing tests, the methods that
failed, and the reasons they failed. A variety of summaries
are provided.

In addition, llilc_checkpass.py can now operate without a
baseline test run. In this case the target run is still
analyzed and failures reported.