Exclude a couple of tests that use vararg methods.
This change also modifies form of exclusions in the exclusion.targets
file.  Because of a recent change to the testing infrastructure each
test project will have its binary and cmd in a distinct directory even
if its source was in the same directory as other test projects. Thus
to exclude a test it is sufficient to give the path to the directory
contaning its cmd file, followed by "\*". The root name of the cmd
file will be the same as the subdirectory name so there is no need to
repeat it.