Emit DebugLocInfo/DebugFileInfo in ObjectWriter for Windows
This basically emits two tables.
One is for file name tables (along with index table).
The other is for native offset to source location info.
Borrowed CodeView emission logic from LLVM mainline. Ideally this code should be refactored to be shared.

Nuget package version (and name for linux) is updated.