Remove dubious usage of getNextNode
LLVM change r252694 (git commit 6426aea5) fixed a problem with getNextNode
when called on the last item in a list (instr in a block / block in a
functino), so that it now returns nullptr as expected, whereas previously
it cast the end iterator sentinel to a bogus node pointer.

This caused some tests to fail in createTemporary where we had evidently
(perhaps unwittingly) worked around the previous behavior by comparing the
result against the end iterator rather than nullptr.

This change updates createTemporary to deal directly in iterators when
setting the insertion point, rather than translating them back to
instructinos.  I audited all the other calls to getNextNode() in LLILC
sources, and changed several of them to likewise deal in iterators, to
simplify reasoning about the code.  There remaining 6 calls all appear to
be correct in that the input is never null/end and null is the desired
output if the input was the last element.