Always ask for the any entry point for a method
We were seeing two different entry point addresses for some methods.

LLILC will insert an explicit null check for calls in the cases that require one, so always request the "any" entry point for a method. Otherwise we need to potentially) keep track of the two entry point addresses for each function, and make sure the right one is used at each call site.

This closes #957 and should also fix #951.