Suppress error popups from asserts
We are seeing timeouts in the lab that seem likely to be caused by asserts raising error dialogs. Set COMPLUS_NoGuiOnAssert in our test environment and use this within LLILC to tell LLVM to generate a stack trace on assert, and tell the CRT we want all assert output to go to stderr instead.

Closes #584.