Add exclusions for tests about to be added to Coreclr.
A new batch of tests is about to be added to CoreClr.
This change adds exclusions for those that fail on
LLILC even though they pass on RyuJit.
Tests that fail without EH enabled for LLILC but that
pass with EH enabled are given issue #13.
The remainder are given issue #970 for later
triage.