Use operand bundles for gc-transition arguments
The prior approach of eagerly generating gc.statepoint calls with
gc-transition arguments is being deprecated (see e.g. r258551), in favor
of attaching these as operand bundles, which should make the semantics
less opaque during optimization.

Since the lowering code that looks for the bundles is in
RewriteStatepointsForGC, we now need to run that pass even with
conservative GC, in methods that have gc transition bundles.  Add a
post-pass to mark all other calls "gc-leaf-function" in that configuration
so that RewriteStatepointsForGC will leave them alone.  Expose a getter on
GenIR to indicate whether we had to generate a managed frame so that we
can avoid this in the common case.

Regardless of GC flavor, inline asm calls (which we use for nops) now must
be explicitly annotated "gc-leaf-function" to avoid lowering trying to
process them as statepoints.