COMDAT sections for ELF
Add support for COMDAT sections for ELF targets. When `ComdatName` is
non-null in CreateCustomSection, create a group whose symbol is
`ComdatName` and associate the created group with it. Add
`ELF::SHF_GROUP` flag to the created section so it will be correctly
coalesced with sections from the same group name when linked to other
object files.

Fix a couple small issues found with my Ubuntu VM's stricter compiler:
disambigurate a type and parameter name and add `const` where it was
being stripped unnecessarily.