Extend SwitchSection API
Provide more precise control over custom sections. Previously we would
create a custom section for Comdat data and refer to it by name when
switching to it. However we really want custom sections with the same
name as existing sections, differing only on attributes or comdat name.
To accomplish this, remove the name -> MCSection mapping and provide an
extended API for switching sections that allows Comdat / attributes to
be specified.